requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromaDB
faiss-cpu
langchain_google_genai
beautifulsoup4
transformers
pip
langchain-community
transformers
